Removed @deephaven/react-hooks dependency
#1469
