+++ packages/jsapi-utils/package.json
@@ -24,7 +24,6 @@
     "@deephaven/filters": "file:../filters",
     "@deephaven/jsapi-types": "file:../jsapi-types",
     "@deephaven/log": "file:../log",
-    "@deephaven/react-hooks": "file:../react-hooks",
     "@deephaven/utils": "file:../utils",
     "lodash.clamp": "^4.0.3",
     "shortid": "^2.2.16"
